# Substantially (~20-45x) faster unique insertion using unique index