@brandur 2024-07-13
> Obviously I love the perf gains, but I'm a little uncertain whether this approach is fully fleshed out. For example, what happens when jobs change states by another mechanism besides the completer (such as the rescuer, or a job cancellation update)? Won't they potentially be left in an inconsistent state with a unique key left in place potentially incorrectly? And what about when a job is manually retried, what happens to the unique key there?

Cool, I think we covered most of this sync, but LMK if you still have concerns about it.

I've since updated the job cancellation paths. This should be ready for review now.