@brandur 2024-07-15
Actually, one tweak I made here is to switch from SHA-3 to SHA-2 because SHA-2 has the nice property of being available in the stdlibs of all of Go, Ruby, and Python, and since we're not using this for security or anything, we don't need the exact state of the art in cryptographic robustness. Even in Go, SHA-3 is still in the /x/ tree, so this is better. I was using the 256-bit variant either way, so plenty of hash space to work with.